removed memory clearing
The feature is not actually doing anything:

1) Resize before memset is causing a memset of length 0, which literally
won't do anything and &in[0] should instead be in.data()

2) Modifying the std::string::data() contents is an undefined behaviour.

3) There are at least two more places that need a cleanup: QString in
QLineEdit widgets and std::string output in std::string
SiteButton::toSave(); Widgets will be hard to analyse and the output
string is being concatenated many times over, likely causing a
reallocation with old data certainly not zeroed.

Implementing it properly will be too complicated and bug prone for this
project and therefore the feature is scrapped.